directory name changed, Respository -> repository
