--- src/main/java/com/test/webtest/domain/airecommendation/Respository/AiRecommendationRepository.java
@@ -1,4 +1,0 @@
-package com.test.webtest.domain.airecommendation.Respository;
-
-public class AiRecommendationRepository {
-}
