+++ src/main/java/com/test/webtest/domain/airecommendation/repository/AiRecommendationRepository.java
@@ -1,0 +1,4 @@
+package com.test.webtest.domain.airecommendation.repository;
+
+public class AiRecommendationRepository {
+}
