--- src/main/java/com/test/webtest/domain/ip/Respository/IpRepository.java
@@ -1,4 +1,0 @@
-package com.test.webtest.domain.ip.Respository;
-
-public class IpRepository {
-}
