+++ src/main/java/com/test/webtest/domain/ip/repository/IpRepository.java
@@ -1,0 +1,4 @@
+package com.test.webtest.domain.ip.repository;
+
+public class IpRepository {
+}
