--- src/main/java/com/test/webtest/domain/priorities/Respository/PrioritiesRepository.java
@@ -1,4 +1,0 @@
-package com.test.webtest.domain.priorities.Respository;
-
-public class PrioritiesRepository {
-}
