+++ src/main/java/com/test/webtest/domain/priorities/repository/PrioritiesRepository.java
@@ -1,0 +1,4 @@
+package com.test.webtest.domain.priorities.repository;
+
+public class PrioritiesRepository {
+}
