--- src/main/java/com/test/webtest/domain/scores/Respository/ScoresRepository.java
@@ -1,4 +1,0 @@
-package com.test.webtest.domain.scores.Respository;
-
-public class ScoresRepository {
-}
