+++ src/main/java/com/test/webtest/domain/scores/repository/ScoresRepository.java
@@ -1,0 +1,4 @@
+package com.test.webtest.domain.scores.repository;
+
+public class ScoresRepository {
+}
