--- src/main/java/com/test/webtest/domain/securityvitals/Respository/SecurityVitalsRepository.java
@@ -1,4 +1,0 @@
-package com.test.webtest.domain.securityvitals.Respository;
-
-public class SecurityVitalsRepository {
-}
