+++ src/main/java/com/test/webtest/domain/securityvitals/repository/SecurityVitalsRepository.java
@@ -1,0 +1,4 @@
+package com.test.webtest.domain.securityvitals.repository;
+
+public class SecurityVitalsRepository {
+}
