--- src/main/java/com/test/webtest/domain/test/Respository/TestRepository.java
@@ -1,4 +1,0 @@
-package com.test.webtest.domain.test.Respository;
-
-public class TestRepository {
-}
