+++ src/main/java/com/test/webtest/domain/test/repository/TestRepository.java
@@ -1,0 +1,4 @@
+package com.test.webtest.domain.test.repository;
+
+public class TestRepository {
+}
