--- src/main/java/com/test/webtest/domain/webvitals/Respository/WebVitalsRepository.java
@@ -1,4 +1,0 @@
-package com.test.webtest.domain.webvitals.Respository;
-
-public class WebVitalsRepository {
-}
