+++ src/main/java/com/test/webtest/domain/webvitals/repository/WebVitalsRepository.java
@@ -1,0 +1,4 @@
+package com.test.webtest.domain.webvitals.repository;
+
+public class WebVitalsRepository {
+}
